Initial
Can append to log file but cannot read. No schema is currently enforced
or represented in the database files and the contents are whatever the
user writes in the cli. Atomicity is not enforced but it is possible to
see whether a write operation is completed properly through log.db.