Modified Project File Structure
Broke main.c down into more files for easier management. This was done
by moving source files into src/ and log and db files into files/. Also
implemented changes to source files so that accessed files are found
using appropriate file paths.

WARNING: From this point forward the binary is going to need to be run
from src/ so that the log and db files will be accessible through a
relative path. This is subject to change.